Refactor: Split connection logic into a few separate crates.
