+++ components/app_conn/src/lib.rs
@@ -1,0 +1,13 @@
+#![crate_type = "lib"]
+#![deny(trivial_numeric_casts, warnings)]
+#![allow(intra_doc_link_resolution_failure)]
+#![allow(
+    clippy::too_many_arguments,
+    clippy::implicit_hasher,
+    clippy::module_inception,
+    clippy::new_without_default
+)]
+
+mod connect;
+
+pub use self::connect::{app_connect_to_node, AppConnectError};
