+++ components/connection/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "offst-app"
+name = "offst-connection"
 version = "0.1.0"
 authors = ["real <real@freedomlayer.org>"]
 license = "MIT OR Apache-2.0"
@@ -14,8 +14,10 @@ timer = { path = "../timer", version = "0.1.0" , package = "offst-timer" }
 proto = { path = "../proto", version = "0.1.0" , package = "offst-proto" }
 signature = { path = "../signature", version = "0.1.0" , package = "offst-signature" }
 net = { path = "../net", version = "0.1.0" , package = "offst-net" }
-app_conn = { path = "../app_conn", version = "0.1.0" , package = "offst-app-conn" }
-connection = { path = "../connection", version = "0.1.0" , package = "offst-connection" }
+keepalive = { path = "../keepalive", version="0.1.0", package = "offst-keepalive" }
+secure_channel = { path = "../secure_channel", version="0.1.0", package = "offst-secure-channel" }
+version = { path = "../version", version="0.1.0", package = "offst-version" }
+
 
 log = "0.4"
 simple_logger = "1.0.1"
