+++ components/connection/src/lib.rs
@@ -1,0 +1,14 @@
+#![deny(trivial_numeric_casts, warnings)]
+#![allow(intra_doc_link_resolution_failure)]
+#![allow(
+    clippy::too_many_arguments,
+    clippy::implicit_hasher,
+    clippy::module_inception,
+    clippy::new_without_default
+)]
+
+extern crate log;
+
+mod secure_connector;
+
+pub use self::secure_connector::create_secure_connector;
